Fix broken get_skde_params test
