+++ data/authority/skde_params.json
@@ -1,10 +1,76 @@
 {
   "params": {
-    "n": "9390473419375365529363714617429845340904444576082356459612604536251158782697317951017301251749869506238465679308829749219246987627651784863828034970978163798243016447675835653547684245161728150895450866183789342351777858067254147531820304840226138781850671395553370058333008429730120412461272681095028484267505241297150696682471665413945257042082408001787636814192425618672587521244549707276380250494111543817379517747294395808094960263641625088443560528887304357921370860984649460079325696888840471911821653957959446501429650644747892630324661347387998311839321964412991855419921220564432328942010888709769594249669",
+    "n": "13909127358298384208582907292993401776729912622203466846326789232728910023510043169191250176550883975111780974910353843994018940992537895074208099104474699608924711221768373426651501565854926084237306768111889066197632264663467214275362215147209207461201495311163198069696839894798783696902033891158632330819972007785156407475737030489596688059035499665735075124493336894285769817199856813170263296274112233747691586762164949966900144659636813294582471103816782444628854673426017017451612313657581797707364645450130653410633803457679151101661658018645737391825509402150641123122485681889229820991842744710961412516281",
     "g": "4",
     "t": 4,
     "h": "4294967296",
     "max_sequencer_number": "2"
   },
-  "signature": "0xe6c88ce4ba6d1dd9a9de4db124fe1710adfa3712ceeb80c4369b76d5bf44260e70b6916ff25f32ebb5f65cf4b0be90974ea226f0529f9a8d38cd365c14ff5cc71b"
+  "signature": [
+    31,
+    37,
+    125,
+    5,
+    20,
+    252,
+    91,
+    33,
+    205,
+    253,
+    103,
+    226,
+    216,
+    162,
+    173,
+    160,
+    165,
+    71,
+    161,
+    27,
+    143,
+    90,
+    221,
+    233,
+    239,
+    42,
+    41,
+    86,
+    143,
+    250,
+    251,
+    120,
+    73,
+    214,
+    86,
+    247,
+    141,
+    157,
+    224,
+    105,
+    185,
+    44,
+    18,
+    226,
+    243,
+    4,
+    170,
+    36,
+    155,
+    229,
+    87,
+    245,
+    223,
+    22,
+    224,
+    189,
+    193,
+    14,
+    39,
+    51,
+    108,
+    5,
+    248,
+    233,
+    27
+  ]
 }
