+++ src/rpc/external/get_latest_session_id.rs
@@ -18,10 +18,8 @@ impl RpcParameter<AppState> for GetLatestSessionId {
     async fn handler(self, _context: AppState) -> Result<Self::Response, RpcError> {
         let session_id = SessionId::get()?;
 
-        // loop {
         return Ok(GetLatestSessionIdResponse {
             latest_session_id: session_id,
         });
-        // }
     }
 }
