+++ src/rpc/external/mod.rs
@@ -1,9 +1,9 @@
 mod get_decryption_key;
 mod get_encryption_key;
 mod get_latest_encryption_key;
-mod get_latest_key_id;
+mod get_latest_session_id;
 
 pub use get_decryption_key::*;
 pub use get_encryption_key::*;
 pub use get_latest_encryption_key::*;
-pub use get_latest_key_id::*;
+pub use get_latest_session_id::*;
