+++ src/tests/integration/get_skde_params.rs
@@ -88,9 +88,11 @@ async fn test_get_skde_params() {
 
     // Compare entire JSON objects at once
     assert_eq!(
-        file_json,
+        file_json["params"],
         response_json,
-        "File content and API response do not match.\nFile: {}\nResponse: {}",
+        "File content and API response do not match.
+        Please re-run test after removing data/authority/skde_params.json and generate new one.
+        File: {}\nResponse: {}",
         serde_json::to_string_pretty(&file_json).unwrap(),
         serde_json::to_string_pretty(&response_json).unwrap()
     );
